# [AArch64] Improve scalar and Neon popcount with SVE CNT.